Migrating prefix_list to prefix_set.
