+++ zebra-rs/src/policy/mod.rs
@@ -10,15 +10,10 @@ pub mod regex;
 
 pub mod com_list;
 
-pub mod plist_ipv4;
-pub use plist_ipv4::{PrefixListIpv4, PrefixListIpv4Map};
-
-pub mod plist_ipv4_config;
-pub use plist_ipv4_config::{prefix_ipv4_commit, prefix_ipv4_exec};
-
-pub mod prefix_set;
-
-pub mod show;
-
 pub mod policy_list;
 pub use policy_list::*;
+
+pub mod plist_ipv4;
+pub use plist_ipv4::{PrefixListIpv4, PrefixListIpv4Map, *};
+
+pub mod show;
